test: share streamlit mocks
